Split unit_test.rb into focused test files per component
Extracted 824-line monolithic test file into 9 dedicated test files,
each named after the component it tests. A shared test_helper.rb
provides common setup (SimpleCov, WebMock, ENV vars, constants).
unit_test.rb is now a thin runner that requires all test files.
